GH-31: [refactor] Update directory and file naming
